Revert "remove memory fixes"
This reverts commit bddb16848c25f313d38e2cccf702e4f4d72c42ff.

+++ src/xray/xrCore/xrMemory.h
@@ -107,8 +107,8 @@ extern XRCORE_API	xrMemory	Memory;
 	IC T*		xr_alloc	(u32 count)				{	return  (T*)Memory.mem_alloc(count*sizeof(T));	}
 	template <class T>
 	IC void		xr_free		(T* &P)					{	if (P) { Memory.mem_free((void*)P); P=NULL;	};	}
-	IC void*	xr_malloc	(u32 size)			{	return	Memory.mem_alloc(size);					}
-	IC void*	xr_realloc	(void* P, u32 size)	{	return Memory.mem_realloc(P,size);				}
+	IC void*	xr_malloc	(size_t size)			{	return	Memory.mem_alloc(size);					}
+	IC void*	xr_realloc	(void* P, size_t size)	{	return Memory.mem_realloc(P,size);				}
 #endif // DEBUG_MEMORY_NAME
 
 XRCORE_API	char* 	xr_strdup	(const char* string);
