+++ src/xray/xrCore/xrMemory_pso_Copy.cpp
@@ -2,13 +2,13 @@
 #pragma hdrstop
 
 #pragma warning(disable:4995)
-void	__stdcall xrMemCopy_x86					(LPVOID dest, const void* src, u32 n)
+void	__stdcall xrMemCopy_x86					(LPVOID dest, const void* src, size_t n)
 {
 	memcpy		(dest,src,n);
 }
 
 #if defined(M_BORLAND) || defined(_M_AMD64)
-void	__stdcall xrMemCopy_MMX					(LPVOID dest, const void* src, u32 n)
+void	__stdcall xrMemCopy_MMX					(LPVOID dest, const void* src, size_t n)
 {
 	memcpy		(dest,src,n);
 }
